backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
requests
python-dotenv
gunicorn
psycopg2-binary
djangorestframework-simplejwt
pytest>=7.0
pytest-django>=4.5
pytest-mock>=3.10
